refactor: split single file into focused modules
- _exceptions.py: exception hierarchy (existing + 3 new)
- _helpers.py: constants, validators, shared argument parsing
- operations.py: do_create, do_list, do_extract and helpers
- cli.py: main(), legacy entry points
- littlefs_tools.py: re-export shim for backward compatibility
